#20 Refactor: using cmd standard
