+++ server/router.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"guzfolio/auth"
